+++ test/test_version.py
@@ -8,4 +8,4 @@ class TestVersion(unittest.TestCase):
         Test the `__version__` global variable.
         """
 
-        self.assertEqual(weaviate.__version__, "3.7.0", "Check if the version is set correctly!")
+        self.assertEqual(weaviate.__version__, "3.8.0", "Check if the version is set correctly!")