+++ weaviate/backup/__init__.py
@@ -1,0 +1,7 @@
+"""
+Module for classifying objects within Weaviate.
+"""
+
+__all__ = ['Backup']
+
+from .backup import Backup